Refactor Asgardeo module types and remove unused augmentation files
