--- packages/nuxt/.npmignore
@@ -1,2 +1,0 @@
-# Ignore TypeScript augmentation files from being processed by consuming apps' build tools
-dist/runtime/types/augments.d.ts
